% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpfa.R
\name{dpfa}
\alias{dpfa}
\title{Write intro here}
\usage{
dpfa(
  data,
  K,
  N,
  M,
  numTime,
  niter = 5000,
  burnin = floor(niter/2),
  init_con = 1,
  init_con_H = 3,
  alpha_psi = 1,
  alpha_phi = 1,
  p0 = 0.5,
  p1 = 0.5,
  rk_a = 100,
  rk_b = 1/rk_a,
  sk_a = 1e+05,
  sk_b = 1/sk_a,
  a0 = 1,
  b0 = 1,
  verbose = T
)
}
\arguments{
\item{data}{Name of the dataset}

\item{K}{The number of classes}

\item{N}{Sample size}

\item{M}{Number of words (variables)}

\item{numTime}{The number of time points}

\item{niter}{fill in}

\item{burnin}{fill in}

\item{init_con}{fill in}

\item{init_con_H}{fill in}

\item{alpha_psi}{fill in}

\item{alpha_phi}{fill in}

\item{p0}{fill in}

\item{p1}{fill in}

\item{rk_a}{fill in}

\item{rk_b}{fill in}

\item{sk_a}{fill in}

\item{sk_b}{fill in}

\item{a0}{fill in}

\item{b0}{fill in}

\item{verbose}{Print progress bar?}
}
\value{
phi_est

psi_est

W_est

theta_est

ZZip_est

ZZip_rowsums
}
\description{
Write intro here
}
\keyword{analysis}
\keyword{dynamic}
\keyword{factor}
\keyword{poisson}
